requirements.txt
# Backend Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
superlinked>=0.1.0
qdrant-client>=1.1.0
openai>=1.0.0
python-dotenv>=0.19.0

# Frontend Dependencies
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.13.0
requests>=2.28.0

# Development Dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.950

# Cloud Dependencies
google-cloud-logging>=3.0.0
google-cloud-run>=0.7.0
google-cloud-storage>=2.0.0 